Fix RUSTSEC-2020-0077 by replacing memmap with memmap2
This replaces the unmaintained memmap crate with the memmap2 crate in
order to fix the recurring security issue.

In order to get a new and shiny build with the updated dependency, we
have to bump the patch version of the package.

Some details here: https://github.com/danburkert/memmap-rs/issues/90